+++ packages/common/CHANGELOG.md
@@ -1,3 +1,32 @@
+# @esri/hub-common [14.200.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.199.0...@esri/hub-common@14.200.0) (2024-09-30)
+
+
+### Features
+
+* **hub-common:** add event scope to catalog ([#1669](https://github.com/Esri/hub.js/issues/1669)) ([9aa595c](https://github.com/Esri/hub.js/commit/9aa595c0131365679eee41174fd012791079a828))
+* user settings org section & notice ([#1655](https://github.com/Esri/hub.js/issues/1655)) ([868fb26](https://github.com/Esri/hub.js/commit/868fb26a432020d6cb3148fca19d4a5f7e678adb))
+
+# @esri/hub-common [14.199.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.198.1...@esri/hub-common@14.199.0) (2024-09-30)
+
+
+### Features
+
+* regate user workspace to qaext and devext so we can try outsi… ([#1660](https://github.com/Esri/hub.js/issues/1660)) ([c634773](https://github.com/Esri/hub.js/commit/c634773a1679a1e8ed55458153cadb5220fe74dc))
+
+## @esri/hub-common [14.198.1](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.198.0...@esri/hub-common@14.198.1) (2024-09-30)
+
+
+### Bug Fixes
+
+* replace hub-radio input with tile-select to resolve a11y bug ([#1665](https://github.com/Esri/hub.js/issues/1665)) ([80dddc0](https://github.com/Esri/hub.js/commit/80dddc0c7e84b35da1c79f4812bc13ca1297f799))
+
+# @esri/hub-common [14.198.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.197.0...@esri/hub-common@14.198.0) (2024-09-27)
+
+
+### Features
+
+* group schema and create group uischema ([#1666](https://github.com/Esri/hub.js/issues/1666)) ([d856c49](https://github.com/Esri/hub.js/commit/d856c494bf5d040c1c77c5f7c526b6d5a27538e4))
+
 # @esri/hub-common [14.197.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.196.1...@esri/hub-common@14.197.0) (2024-09-26)